+++ util/sokol_gl.h
@@ -2905,7 +2905,7 @@ static void _sgl_log(sgl_log_item_t log_item, uint32_t log_level, uint32_t line_
             const char* filename = 0;
             const char* message = 0;
         #endif
-        _sgl.desc.logger.func("sgl", log_level, log_item, message, line_nr, filename, _sgl.desc.logger.user_data);
+        _sgl.desc.logger.func("sgl", log_level, (uint32_t)log_item, message, line_nr, filename, _sgl.desc.logger.user_data);
     } else {
         // for log level PANIC it would be 'undefined behaviour' to continue
         if (log_level == 0) {
